Adapting monsterService
